+++ library/src/main/java/me/denley/notary/FileFilter.java
@@ -1,7 +1,0 @@
-package me.denley.notary;
-
-public interface FileFilter {
-
-    public boolean accept(File file);
-
-}