+++ library/src/main/java/me/denley/notary/SyncableFileFilter.java
@@ -1,0 +1,8 @@
+package me.denley.notary;
+
+public interface SyncableFileFilter {
+
+    public boolean display(File file);
+    public boolean autoSync(File file);
+
+}