+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Directory observer now properly observes only transactions in the source directory
 - Directory observer can now sort files with a custom comparator
 - Directory observer now records sync state with remote device
+- Autosyncable items
 
 ## 0.1.0
 ### Added